// packages/core/src/models/field/core/field.types.enum.ts
/**
 * FieldType enum defines the avaialble Formidable Form Field Types
 */
export enum FieldType {
    Number = 'number',
    Text = 'text',
    Textarea = 'textarea',
    Link = 'link',
    Email = 'email',
    Password = 'password',
    NumberRange = 'numberrange',
    Date = 'date',
    Time = 'time',
    Datetime = 'datetime',
    File = 'file',
    Image = 'image',
    Audio = 'audio',
    Video = 'video'
}